src/backend/requirements.txt
fastapi
uvicorn
pydantic
websockets
aiofiles
huggingface_hub[cli,hf_transfer]
pyyaml
docker